Support oci compatible user setting
This allows the user field to be set according to the oci user spec. It
can be uid, user, uid:gid, user:group, uid,group, user:gid. If named
users or groups are used, then nss handling is automatically enabled.
This also allows supplemental groups to be added in the config and adds
a missing key file.